requirements.txt
Flask
boto3
python-dotenv
supabase
gunicorn
discord